Address editor review findings
- Auto-capitalize sentences in the comment body field (parity with the
  legacy field and the reply box).
- Remove the now-dead InputSettings/inputSettings and write-only
  isFromRegisteredUser state, and the tests that only asserted the
  hardcoded all-enabled flags.
- Extract the duplicated SnackbarMessageHolder->SnackbarHostState bridge
  into a shared showMessage() extension, reused by the editor and detail
  fragments.
- Disable the Save button while saving/loading to prevent a double-submit.
