+++ WordPress/src/main/java/org/wordpress/android/ui/comments/unified/CommentEssentials.kt
@@ -5,8 +5,7 @@ data class CommentEssentials(
     val userName: String = "",
     val commentText: String = "",
     val userUrl: String = "",
-    val userEmail: String = "",
-    val isFromRegisteredUser: Boolean = true
+    val userEmail: String = ""
 ) {
     /**
      * Checks if this instance of CommentEssentials is valid. An invalid instance should not be used to display data.
